requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
Appium-Python-Client==1.2.0
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
cryptography==36.0.1
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==2.4.7
pytest==6.2.4
pytest-rerunfailures==10.0
requests==2.25.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
transliterate==1.10.2
urllib3==1.26.5
